+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Rename `Decider.state` to `Decider.initial_state`
 * Allow to use anything as state
 * Use tuple-like array for composition
+* Do not raise error when deciding unknown command
+* Do not raise error when evolving unknown event
 
 # 0.2.0